Added a test shader for blending the edges of a cube to another texture.
Can be used to make non-sharp edges automatically. Right now nothing is automatic, and you need to set both texture, blend values and direction of blend but all this is possible to do automatically, via e.g. an editor script (or in game).

Can later be extended to animate the blend, to make moving edges (or shores).

Signed-off-by: Kalle Sjöström <c06ksm@gmail.com>